hubble: Add support for SockLB tracing
This commit adds a Hubble parser for `TraceSockNotify` events. Those
events are emitted whenever SockLB is performed, see cilium/cilium#20492
for details about the datapath tracing.

This parser is similar to the L3/L4 event parser. But because the events
are traced on the socket level, we do not know the source endpoint id,
source endpoint IP address or source TCP/UDP port of the event. We only
know the source cgroup ID. From the cgroup, we can derive which local
pod emitted the event (via `GetPodMetadataForContainer`). From the pod
IP we can then perform an IP-based lookup in either the local endpoint
list or the IPCache to determine the source endpoint metadata (endpoint
id, numeric identity, labels, etc). The lookup for destination IP is
almost identical to the threefour parser, with the exception that we do
not have any datapath identity available for fallback.

The resulting flow looks similar to regular `FlowType_L3_L4`, with a
couple of key differences:

 - The cgroup id of the container is available
 - Information if the event is pre or post translation point is
   available
 - The socket cookie is available
 - The TCP/UDP source port is _not_ known
 - TCP flags are _not_ known
 - Ethernet headers are _not_ known

Unfortunately, sock events can also be performed by non-pod entities
(e.g. the host cgroup). In that case, we cannot perform a mapping to any
local endpoint and therefore the source IP and endpoint data are left
empty. The cgroup id may also not be available on older kernels or
systems with the systemd cgroup driver (though this will likely be fixed
in a follow-up PR). This fact that the source IP may be empty/unknown is
also key difference to the flows which have been produced by Hubble so
far.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>